requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
openai
pdfkit
python-dotenv
